+++ src/components/dropdown-filter/dropdown-filter.js
@@ -5,6 +5,7 @@ import I18n from 'i18n-js';
 import classNames from 'classnames';
 import escapeStringRegexp from 'escape-string-regexp';
 import Dropdown from './../dropdown';
+import Link from './../link';
 
 /**
  * A dropdown filter widget.
@@ -108,6 +109,22 @@ class DropdownFilter extends Dropdown {
      * @type {Function}
      */
     create: PropTypes.func,
+
+    /**
+     * Customizes text for the create functionality of the dropdown.
+     *
+     * @property createText
+     * @type {String}
+     */
+    createText: PropTypes.string,
+
+    /**
+     * Customizes the Carbon Icon type for the create functionality of the dropdown.
+     *
+     * @property createIconType
+     * @type {String}
+     */
+    createIconType: PropTypes.string,
 
     /**
      * Should the dropdown act and look like a suggestable input instead.
@@ -328,23 +345,27 @@ class DropdownFilter extends Dropdown {
         html = [original];
 
     if (this.state.open && this.props.create) {
-      let text = 'Create ';
+      let createText = I18n.t('dropdown_filter.create_text', { defaultValue: 'Create' });
 
-      if (this.state.filter) {
-        text += `"${this.state.filter}"`;
+      if (this.props.createText) {
+        createText = this.props.createText;
+      } else if (this.state.filter) {
+        createText += ` "${this.state.filter}"`;
       } else {
-        text += 'New';
+        createText += I18n.t('dropdown_filter.new_text', { defaultValue: ' New' });
       }
 
       html.push(
-        <a
+        <Link
+          icon={ this.props.createIconType || 'add' }
+          iconAlign='left'
           className='carbon-dropdown__action'
           data-element='create'
           key='dropdown-action'
           onClick={ this.handleCreate }
         >
-          { text }
-        </a>
+          { createText }
+        </Link>
       );
     }